Implement coordinator registration backend and remove unused axios client
Coordinator registration was calling five API endpoints that never existed
(/api/states, /districts/by-state, /blocks/by-district, /schools/by-block,
/teachers). Added them, backed by real seeded data via a new shared
geoData.ts (dedupes the state/district master list that seed.ts previously
inlined). Switched coordinatorService.ts off its one-off axios client onto
the app's standard apiFetch, and removed the now-unused axios dependency.
