+++ frontend/package.json
@@ -11,7 +11,6 @@
   },
   "dependencies": {
     "@tanstack/react-query": "^5.101.2",
-    "axios": "^1.18.1",
     "lucide-react": "^0.546.0",
     "motion": "^12.23.24",
     "qrcode": "^1.5.4",
